requirements.txt
beautifulsoup4
requests
pandas
networkx
tenacity
numpy
matplotlib
seaborn
scipy
igraph
plotly
tidyverse
ggplot2
ggraph
reshape2
showtext
patchwork
python-dotenv